feat: scroll to histogram on fetch error and restyle tooltip close button
Two changes that improve feedback on the map tooltip:

- Some coordinates (e.g. 39.05, 56.51) cause the point-data endpoint
  to fail with "The regex parameter does not have a valid EPSG code".
  The inline error renders inside the histogram card, but if the
  sidebar wasn't scrolled to it, the user activates the histogram
  and thinks nothing happened. Add an effect that scrolls to the
  histogram anchor whenever a new (layer, lon/lat) pair surfaces a
  fetch error. The error message also picks up `role="alert"` and a
  `point-histogram-error` testid so screen readers announce it and
  it can be asserted in tests.

- Extract the existing in-tooltip `scrollToHistogram` helper into
  `src/lib/scroll-to-histogram.ts` so both the tooltip and the
  point-histogram component share the same anchor + sticky-offset
  logic.

- Restyle the tooltip close button to match the design spec:
  - 34x34 outer circle, `bg-black-150`, 1px `border-white-500/[0.08]`
    for the subtle outer ring.
  - Positioned `top-5` (20px from the tooltip top) and `right-[-17px]`
    so the cross sits centered on the tooltip's right border.
  - Nested 18x18 inner ring (`border-white-500`) wrapping a 9x9 LuX
    cross at stroke-width 1.5, reproducing the SVG's two-circle look
    with containers instead of inline SVG.
